+++ libraries/search/lib/objective/exception/ExceptionObjectiveFunction.ts
@@ -16,9 +16,8 @@
  * limitations under the License.
  */
 
-import { Encoding } from "../Encoding";
-
-import { ObjectiveFunction } from "./ObjectiveFunction";
+import { Encoding } from "../../Encoding";
+import { ObjectiveFunction } from "../ObjectiveFunction";
 
 /**
  * Objective function for the exception criterion.
